+++ src/addons/meta_human_dna/ui/callbacks.py
@@ -696,10 +696,17 @@ def update_body_rig_bone_group_selection(self, context):
         body.select_bone_group()
 
 def update_face_pose(self, context):
-    from ..utilities import get_active_head
-    head = get_active_head()
-    if head:
-        head.set_face_pose()
+    from ..utilities import get_head
+    active_instance = get_active_rig_logic()
+    if not active_instance:
+        return
+
+    # update all instances with the same face board
+    for instance in context.scene.meta_human_dna.rig_logic_instance_list:
+        if instance.face_board == active_instance.face_board:
+            head = get_head(instance.name)
+            if head:
+                head.set_face_pose()
 
 def update_head_to_body_constraint_influence(self, context):
     from ..utilities import get_active_head